+++ src/items.cpp
@@ -702,7 +702,7 @@ void Items::parseItemNode(const pugi::xml_node& itemNode, uint16_t id)
 				}
 
 				case ITEM_PARSE_ROTATETO: {
-					it.rotateTo = pugi::cast<int32_t>(valueAttribute.value());
+					it.rotateTo = pugi::cast<uint16_t>(valueAttribute.value());
 					break;
 				}
